fix(ScheduleComponent): days without classes do not create empty spaces
filterClasses and sortClasses were refactored from services to pipes. Schedule component does filtering instead of DayColumnComponent
