+++ src/app/fitness-schedule/day-column/day-column.component.ts
@@ -2,9 +2,6 @@ import { Component, Input, OnInit } from '@angular/core';
 import { MappingService } from '../services/mapping.service';
 import { FitnessClassesPerDay } from '../interfaces/fitness-classes-per-day';
 import { FitnessClass } from '../interfaces/fitness-class';
-import { FilterClassesService } from '../services/filter-classes.service';
-import { SortClassesService } from '../services/sort-classes.service';
-import { FilterState } from '../store/filter/filter-state';
 
 @Component({
   selector: 'sfs-day-column',
@@ -14,19 +11,12 @@ import { FilterState } from '../store/filter/filter-state';
 export class DayColumnComponent implements OnInit {
 
   @Input() classesPerDay: FitnessClassesPerDay = {day: null, classes: []};
-  @Input() filter: FilterState = {};
   @Input() removedFavorites = [];
-  filteredClasses: FitnessClass[] = [];
 
-  constructor(private mappingService: MappingService,
-              private filterService: FilterClassesService,
-              private orderService: SortClassesService) {
+  constructor(private mappingService: MappingService) {
   }
 
   ngOnInit() {
-    this.filteredClasses = [this.classesPerDay.classes]
-      .map(classes => this.filterService.filter(classes, this.filter))
-      .map(classes => this.orderService.sort(classes))[0];
   }
 
   getDayName(day) {
