+++ src/app/fitness-schedule/schedule/schedule.component.spec.ts
@@ -7,6 +7,7 @@ import { FlexLayoutModule } from '@angular/flex-layout';
 import { OrderDaysPipe } from '../pipes/order-days.pipe';
 import { ScheduleActions } from '../store/schedule/schedule.actions';
 import { NgReduxTestingModule } from '@angular-redux/store/lib/testing';
+import { FilterClassesPipe } from '../pipes/filter-classes.pipe';
 
 describe('ScheduleComponent', () => {
   let component: ScheduleComponent;
@@ -18,6 +19,7 @@ describe('ScheduleComponent', () => {
         ScheduleComponent,
         MockDayColumnComponent,
         OrderDaysPipe,
+        FilterClassesPipe
       ],
       imports: [
         SfsMaterialModule,
